Set up docker-compose.yml (not working), API and Repository interface
