+++ build.gradle
@@ -32,7 +32,7 @@ dependencies {
 	implementation 'junit:junit:4.12'
 	implementation 'junit:junit:4.12'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	
+
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	annotationProcessor 'org.projectlombok:lombok'
