+++ docker-compose.yml
@@ -1,0 +1,37 @@
+ services:
+
+    build: wonda-cabinet-inc
+    hostname: wonda
+    environment:
+        - SPRING_PROFILES_ACTIVE=docker
+      depends_on:
+        mysql:
+          condition: service_healthy
+
+    mysql:
+      image: mysql:5.7
+      #mem_limit: 350m
+      ports:
+        - "3307:3306"
+      environment:
+        - MYSQL_ROOT_PASSWORD=rootpwd
+        - MYSQL_DATABASE=wondacabinetincdb
+        - MYSQL_USER=user
+        - MYSQL_PASSWORD=pwd
+      volumes:
+        - ./data/mysql1:/var/lib/mysql
+        - ./data/init.d:/docker-entrypoint-initdb.d
+      healthcheck:
+        test:
+          [
+              "CMD",
+              "mysqladmin",
+              "ping",
+              "-uuser",
+              "-ppwd",
+              "-h",
+              "localhost",
+          ]
+        interval: 10s
+        timeout: 5s
+        retries: 10
