+++ src/main/java/com/wondacabinetinc/datalayer/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.wondacabinetinc.datalayer;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+}
