--- src/main/java/com/wondacabinetinc/wondacabinetinc/datalayer/Item.java
@@ -1,39 +1,0 @@
-package com.wondacabinetinc.wondacabinetinc.datalayer;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name="items")
-public class Item {
-
-    @Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int itemId;
-
-    @Column(name="item_name")
-    private String itemName;
-
-    public Item(int itemId, String itemName) {
-        this.itemId = itemId;
-        this.itemName = itemName;
-    }
-
-    public Item() {
-    }
-
-    public int getItemId() {
-        return itemId;
-    }
-
-    public void setItemId(int itemId) {
-        this.itemId = itemId;
-    }
-
-    public String getItemName() {
-        return itemName;
-    }
-
-    public void setItemName(String itemName) {
-        this.itemName = itemName;
-    }
-}
