+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
-
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/WondaCabinetIncDb
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/WondaCabinetIncDb?useSSL=false
 spring.datasource.username=user
 spring.datasource.password=pwd
 spring.datasource.driver-class-name =com.mysql.cj.jdbc.Driver
-#spring.jpa.show-sql: true
+#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.show-sql:true
