+++ src/main/resources/application.yml
@@ -13,12 +13,12 @@
 #logging:
 #  level:
 #    root: INFO
-#    com.liang: DEBUG
+#    com.wondacabinetinc: DEBUG
 #    org.hibernate.SQL: DEBUG
 #    org.hibernate.type.descriptor.sql.BasicBinder: TRACE
 #
 #---
-##spring.profiles: docker
+#spring.profiles: docker
 #
 #spring.datasource:
 #  url: jdbc:mysql://mysql/WondaCabinetIncDb
