+++ src/main/resources/db/mysql/data.sql
@@ -1,3 +1,3 @@
 INSERT IGNORE INTO orders VALUES(1, 'Done', 123321, 'Design');
-INSERT IGNORE INTO orders VALUES(2, 'Done', 123321, 'Design');
-INSERT IGNORE INTO orders VALUES(3, 'Done', 123321, null);
+INSERT IGNORE INTO orders VALUES(2, 'Done', 456647, 'Design');
+INSERT IGNORE INTO orders VALUES(3, 'Done', 789789, null);
