--- src/test/java/com/wondacabinetinc/wondacabinetinc/WondacabinetincApplicationTests.java
@@ -1,13 +1,0 @@
-package com.wondacabinetinc.wondacabinetinc;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WondacabinetincApplicationTests {
-
-/*	@Test
-	void contextLoads() {
-	}*/
-
-}
